(fix) moved cabin option to FareInformativeBestPricingWithoutPnrOptions
