+++ src/Amadeus/Client/Struct/Fare/PricePNRWithBookingClass12.php
@@ -172,10 +172,6 @@ class PricePNRWithBookingClass12 extends BaseWsMessage
         if (!empty($options->overrideOptionsWithCriteria)) {
             throw new OptionNotSupportedException('Override Options With Criteria are not supported in version 12 or lower');
         }
-
-        if (!empty($options->cabin)) {
-            throw new OptionNotSupportedException('Cabin option is not supported in version 12 or lower');
-        }
     }
 
     /**
