+++ src/Amadeus/Client/Struct/Fare/PricePNRWithBookingClass13.php
@@ -24,7 +24,6 @@ namespace Amadeus\Client\Struct\Fare;
 
 use Amadeus\Client\RequestCreator\MessageVersionUnsupportedException;
 use Amadeus\Client\RequestOptions\Fare\PricePnr\AwardPricing;
-use Amadeus\Client\RequestOptions\Fare\PricePnr\Cabin;
 use Amadeus\Client\RequestOptions\Fare\PricePnr\ExemptTax;
 use Amadeus\Client\RequestOptions\Fare\PricePnr\FareBasis;
 use Amadeus\Client\RequestOptions\Fare\PricePnr\FormOfPayment;
@@ -177,11 +176,6 @@ class PricePNRWithBookingClass13 extends BasePricingMessage
             self::makeOverrideOptionsWithCriteria($options->overrideOptionsWithCriteria, $priceOptions)
         );
 
-        $priceOptions = self::mergeOptions(
-            $priceOptions,
-            self::loadCabin($options->cabin)
-        );
-
         // All options processed, no options found:
         if (empty($priceOptions)) {
             $priceOptions[] = new PricingOptionGroup(PricingOptionKey::OPTION_NO_OPTION);
@@ -264,20 +258,6 @@ class PricePNRWithBookingClass13 extends BasePricingMessage
         return $opt;
     }
 
-    /**
-     * @param Cabin $cabin
-     * @return PricePnr13\PricingOptionGroup[]
-     */
-    protected static function loadCabin($cabin)
-    {
-        $opt = [];
-
-        if ($cabin instanceof Cabin) {
-            $opt[] = new PricingOptionGroup(PricingOptionKey::OPTION_CABIN, $cabin->criteriaDetails);
-        }
-
-        return $opt;
-    }
 
     /**
      * @param FareBasis[] $pricingsFareBasis
